Add new Js file and fix edit error
