+++ webpack.config.js
@@ -3,7 +3,7 @@ const HtmlWebpackPlugin = require('html-webpack-plugin');
 
 module.exports = {
   mode: 'development',
-  entry: path.resolve(__dirname, 'src/index.js'),
+  entry: path.resolve(__dirname, 'src/functionality.js'),
   output: {
     path: path.resolve(__dirname, 'dist'),
     filename: 'bundle.js',
